Revert "renamed index to server"
