+++ dummyBackend/api/index.js
@@ -1,0 +1,25 @@
+const jsonServer = require('json-server')
+const cors = require('cors')
+const path = require('path')
+
+const server = jsonServer.create()
+const router = jsonServer.router('db.json')
+const middlewares = jsonServer.defaults()
+
+server.use(cors())
+server.use(jsonServer.bodyParser)
+server.use(middlewares)
+
+server.use(jsonServer.rewriter({
+  '/api/*': '/$1',
+  '/blog/:resource/:id/show': '/:resource/:id'
+}))
+server.use(router)
+
+const PORT = 8000
+
+server.listen(PORT, () => {
+  console.log(`JSON Server is running on http://localhost:${PORT}`)
+})
+
+module.exports = server
