+++ dummyBackend/package.json
@@ -2,9 +2,9 @@
   "name": "dummyBackend",
   "version": "1.0.0",
   "description": "",
-  "main": "api/server.js",
+  "main": "api/index.js",
   "scripts": {
-    "start": "node api/server.js"
+    "start": "node api/index.js"
   },
   "keywords": [],
   "author": "",
